backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
python-decouple==3.8
psycopg2-binary==2.9.9
gunicorn==22.0.0
whitenoise==6.8.1